#192: extract sendOne into named PeerFileSender (data layer) — keep BatchSender transport-agnostic
